Move controls out of subblocks
